+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased [3.4.0-SNAPSHOT]
+### Added
+### Changed
+### Fixed
+
 ## [3.3.0] - 2018-4-24
 ### Added
  - Check for spaces in output path for KtLint versions earlier