+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
   - KtLint versions prior to 0.10.0 are not supported anymore
   - Gradle versions prior to 4.3 are not supported anymore
+  - Deprecated ReporterType typealias
+  - Deprecated reporter field from extension
 
 ## [3.3.0] - 2018-4-24
 ### Added